proyekt tugash arafasida buglar tuzatildi

+++ models/category.js
@@ -2,7 +2,7 @@ const { DataTypes } = require('sequelize');
 const {db} = require('../config/db');
 
 const Category = db.define('Category', {
-  id: { type: DataTypes.BIGINT, primaryKey: true, autoIncrement: true },
+  id: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
   name: { type: DataTypes.STRING, allowNull: false }
 }, {
   timestamps: true
